+++ src/img/CHANGES.md
@@ -1,5 +1,13 @@
 # imgadm changelog
 
+## 2.5.0
+
+- `imgadm create -s <prepare-image-script> ...` support to automatic running
+  of a preparation script, gated by VM snapshotting and rollback for safety.
+  This makes image creation easier (fewer steps, do not need to manually
+  prepare and stop the VM) and safer (snapshot/rollback ensures an unchanged
+  and usable original VM on completion).
+
 ## 2.4.0
 
 - Add `imgadm update [UUID...]` support for updating specific image uuids.